+++ Areas/Front/ViewModels/Page/PageTreeVM.cs
@@ -1,15 +1,9 @@
-﻿using Bonsai.Areas.Front.ViewModels.Tree;
-
-namespace Bonsai.Areas.Front.ViewModels.Page
+﻿namespace Bonsai.Areas.Front.ViewModels.Page
 {
     /// <summary>
-    /// Elements of the page's tree.
+    /// Blank class for tree page VM.
     /// </summary>
     public class PageTreeVM: PageTitleVM
     {
-        /// <summary>
-        /// Tree data.
-        /// </summary>
-        public TreeVM Tree { get; set; }
     }
 }
