+++ Areas/Front/Views/Page/Tree.cshtml
@@ -1,5 +1,4 @@
 ﻿@using Bonsai.Areas.Front.ViewModels.Page
-@using Newtonsoft.Json
 @model PageVM<PageTreeVM>
 @{
     Layout = "~/Areas/Front/Views/Page/PageLayout.cshtml";
@@ -9,7 +8,7 @@
 }
 
 <div class="tree">
-    <div class="tree-wrapper" data-src="#tree-contents">
+    <div class="tree-wrapper" data-key="@Model.Body.Key" data-root="@Model.Body.Id">
         <svg class="tree-back" :width="width" :height="height">
             <polyline :points="e.points" v-for="e in edges" />
         </svg>
@@ -27,10 +26,6 @@
         </h4>
     </div>
 </div>
-
-<script id="tree-contents" type="application/json">
-    @Html.Raw(JsonConvert.SerializeObject(Model.Body.Tree, Formatting.Indented))
-</script>
 
 <script id="tree-card-template" type="text/x-template">
     <div :class="{'tree-card-wrapper': true, male: value.info.IsMale, female: !value.info.IsMale, deceased: value.info.IsDead, active: active}"
@@ -52,6 +47,5 @@
 </script>
 
 @section Scripts {
-    <script src="@Url.Content("~/assets/scripts/vendor-elk.js")"></script>
     <script src="@Url.Content("~/assets/scripts/vendor-vue.js")"></script>
 }
